Update clientlog package to use new logging package log.go
Also remove test, firstly the new logger points to a new destination
to store/output logs which cannot be intercepted (or at least I cannot
get it to do so). Secondly the tests are checking the data is added to
the logged event, this appears to test how the logger works which in essense
is tested in the log.go package itself.