Add population type field to Job model
We need this in order to determine the underlying Cantabular dataset to
pass to `GetGeographyDimensions`. While it's possible to determine with
an additional API call to the Dataset API, since the data is already
here we may as well avoid a request.

Possibly the testing strategy should be revisited, since ideally we
would just test for deep equality between a stub `Model` and the
deserialized response, but that would be a departure from the
existing method's tests (feels overkill for one field addition).

Resolves: 5569 (partial)