Add 'SubmitFilter' function
A new endpoint is required in dp-cantabular-filter-flex-api to submit
the user choices to generate the cantabular downloads, so we need to
create an associated function in dp-api-clients-go.

A new function should be created to handle the request to submit a
filter for a cantabular dataset which should POST to
`/filters/{filterid}/submit` in `dp-cantabular-filter-flex-api`.

Resolves: 5614